Rename test→json
